+++ src/Fable.Transforms/BabelPrinter.fs
@@ -650,9 +650,8 @@ module PrinterExtensions =
                     printer.PrintNewLine()
 
                     for line in lines do
-#if !FABLE_COMPILER
-                        let line = System.Web.HttpUtility.HtmlDecode(line)
-#endif
+                        let line = Naming.xmlDecode line
+
                         printer.Print(" * ")
                         printer.Print(line.Trim())
                         printer.PrintNewLine()