+++ src/Fable.Transforms/Replacements.fs
@@ -2123,7 +2123,7 @@ let nullables (com: ICompiler) (_: Context) r (t: Type) (i: CallInfo) (thisArg: 
     | "get_HasValue", Some c -> Test(c, OptionTest true, r) |> Some
     | _ -> None
 
-// See fable-library/Option.ts for more info on how options behave in Fable runtime
+// See fable-library-ts/Option.ts for more info on how options behave in Fable runtime
 let options isStruct (com: ICompiler) (_: Context) r (t: Type) (i: CallInfo) (thisArg: Expr option) (args: Expr list) =
     match i.CompiledName, thisArg with
     | "Some", _ -> NewOption(List.tryHead args, t.Generics.Head, isStruct) |> makeValue r |> Some
@@ -2636,6 +2636,60 @@ let dictionaries (com: ICompiler) (ctx: Context) r t (i: CallInfo) (thisArg: Exp
       Some c -> Helper.InstanceCall(c, methName, t, args, i.SignatureArgTypes, ?loc = r) |> Some
     | _ -> None
 
+let conditionalWeakTable (com: ICompiler) (ctx: Context) r t (i: CallInfo) (thisArg: Expr option) (args: Expr list) =
+    match i.CompiledName, thisArg with
+    | ".ctor", _ ->
+        match i.GenericArgs with
+        | [ keyType; _ ] ->
+            match keyType with
+            | Boolean
+            | String
+            | Number _ ->
+                $"ConditionalWeakTable does not support primitive keys in JS"
+                |> addError com ctx.InlinePath r
+            | _ -> ()
+        | _ ->
+            $"Unexpected number of generic arguments for ConditionalWeakTable: %A{i.GenericArgs}"
+            |> addError com ctx.InlinePath r
+
+        Helper.LibCall(
+            com,
+            "ConditionalWeakTable",
+            "default",
+            t,
+            args,
+            i.SignatureArgTypes,
+            isConstructor = true,
+            genArgs = i.GenericArgs,
+            ?loc = r
+        )
+        |> Some
+    | "Add", _ ->
+        Helper.LibCall(com, "MapUtil", "addToDict", t, args, i.SignatureArgTypes, ?thisArg = thisArg, ?loc = r)
+        |> Some
+    | "GetOrCreateValue", _ -> None
+    | "GetValue", _ ->
+        Helper.LibCall(
+            com,
+            "MapUtil",
+            "getItemFromDictOrCreate",
+            t,
+            args,
+            i.SignatureArgTypes,
+            ?thisArg = thisArg,
+            ?loc = r
+        )
+        |> Some
+    | "TryAdd", _ ->
+        Helper.LibCall(com, "MapUtil", "tryAddToDict", t, args, i.SignatureArgTypes, ?thisArg = thisArg, ?loc = r)
+        |> Some
+    | "TryGetValue", _ ->
+        Helper.LibCall(com, "MapUtil", "tryGetValue", t, args, i.SignatureArgTypes, ?thisArg = thisArg, ?loc = r)
+        |> Some
+    | ReplaceName [ "AddOrUpdate", "set"; "Clear", "clear"; "Remove", "delete" ] meth, Some c ->
+        Helper.InstanceCall(c, meth, t, args, i.SignatureArgTypes, ?loc = r) |> Some
+    | _ -> None
+
 let hashSets (com: ICompiler) (ctx: Context) r t (i: CallInfo) (thisArg: Expr option) (args: Expr list) =
     match i.CompiledName, thisArg, args with
     | ".ctor", _, _ ->
@@ -2863,12 +2917,12 @@ let dateTime (com: ICompiler) (ctx: Context) r t (i: CallInfo) (thisArg: Expr op
 
             // JavaScript Date doesn't support microseconds precision
             | 8, Number(Int32, NumberInfo.Empty) ->
-                "JavaScript Date with doesn't support microseconds precision"
+                "JavaScript Date doesn't support microseconds precision"
                 |> addError com ctx.InlinePath r
 
                 None
             // | 8, Number(_, NumberInfo.IsEnum ent) when ent.FullName = "System.DateTimeKind" ->
-            //     "JavaScript Date with doesn't support microseconds precision" |> addError com ctx.InlinePath r
+            //     "JavaScript Date doesn't support microseconds precision" |> addError com ctx.InlinePath r
             //     None
 
             | _ ->
@@ -2905,23 +2959,20 @@ let dateTimeOffset (com: ICompiler) (ctx: Context) r t (i: CallInfo) (thisArg: E
         | ExprType(Number(Int64, _)) :: _ ->
             Helper.LibCall(com, moduleName, "fromTicks", t, args, i.SignatureArgTypes, ?loc = r)
             |> Some
-        | ExprType(DeclaredType(e, [])) :: _ when e.FullName = Types.datetime ->
+        | ExprType(DeclaredType(ent, [])) :: _ when ent.FullName = Types.datetime ->
             Helper.LibCall(com, moduleName, "fromDate", t, args, i.SignatureArgTypes, ?loc = r)
             |> Some
         | _ ->
-            let last = List.last args
-
-            match args.Length, last.Type with
-            | 7, Number(_, NumberInfo.IsEnum ent) when ent.FullName = "System.DateTimeKind" ->
-                let args = (List.take 6 args) @ [ makeIntConst 0; last ]
-
-                let argTypes = (List.take 6 i.SignatureArgTypes) @ [ Int32.Number; last.Type ]
-
-                Helper.LibCall(com, "Date", "create", t, args, argTypes, ?loc = r) |> Some
-
-            | _ ->
+            match args.Length with
+            | 7
+            | 8 ->
                 Helper.LibCall(com, moduleName, "create", t, args, i.SignatureArgTypes, ?loc = r)
                 |> Some
+            | _ ->
+                "JavaScript Date doesn't support microseconds precision"
+                |> addError com ctx.InlinePath r
+
+                None
     | "ToString" ->
         Helper.LibCall(com, "Date", "toString", t, args, i.SignatureArgTypes, ?thisArg = thisArg, ?loc = r)
         |> Some
@@ -3864,6 +3915,7 @@ let private replacedModules =
             Types.dictionary, dictionaries
             Types.idictionary, dictionaries
             Types.ireadonlydictionary, dictionaries
+            Types.conditionalWeakTable, conditionalWeakTable
             Types.ienumerableGeneric, enumerables
             Types.ienumerable, enumerables
             Types.valueCollection, enumerables