+++ src/Fable.Transforms/Transforms.Util.fs
@@ -331,6 +331,9 @@ module Types =
     let ireadonlydictionary = "System.Collections.Generic.IReadOnlyDictionary`2"
 
     [<Literal>]
+    let conditionalWeakTable = "System.Runtime.CompilerServices.ConditionalWeakTable`2"
+
+    [<Literal>]
     let hashset = "System.Collections.Generic.HashSet`1"
 
     [<Literal>]
@@ -884,7 +887,7 @@ module AST =
         | _ -> true
 
     /// For unit, unresolved generics or nested options or unknown types,
-    /// create a runtime wrapper. See fable-library/Option.ts for more info.
+    /// create a runtime wrapper. See fable-library-ts/Option.ts for more info.
     let rec mustWrapOption =
         function
         | Any