+++ src/fable-standalone/CHANGELOG.md
@@ -7,6 +7,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 1.5.0 - 2024-03-20
+
+### Changed
+
+* Fable 4.17.0
+
+## 1.4.0 - 2024-03-20
+
+### Changed
+
+* Fable 4.16.0
+
+## 1.3.0 - 2024-03-18
+
+### Changed
+
+* Fable 4.15.0
+
 ## 1.2.0 - 2024-03-01
 
 ### Changed